srcs/Back/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.2.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart
sqlalchemy
jinja2
asyncpg
uvicorn[standard]
email-validator
PyJWT